+++ core/src/lib.rs
@@ -10,6 +10,7 @@
 #![deny(missing_docs)]
 #![no_std]
 
+mod addresses;
 mod encryption;
 mod error;
 mod keys;
@@ -18,17 +19,39 @@ mod note;
 #[cfg(feature = "alloc")]
 mod transaction;
 
+pub use addresses::stealth::StealthAddress;
+pub use addresses::sync::SyncAddress;
+pub use addresses::Ownable;
 pub use encryption::aes;
 pub use error::Error;
 pub use keys::hash;
 pub use keys::public::PublicKey;
 pub use keys::secret::SecretKey;
-pub use keys::stealth::StealthAddress;
-pub use keys::sync::SyncAddress;
 pub use keys::view::ViewKey;
-pub use keys::Ownable;
-pub use note::{Note, NoteType};
+pub use note::{Note, NoteType, ENCRYPTION_SIZE as NOTE_ENCRYPTION_SIZE};
 
 #[cfg(feature = "alloc")]
 /// Transaction Skeleton used by the phoenix transaction model
 pub use transaction::TxSkeleton;
+
+use dusk_jubjub::{
+    JubJubAffine, JubJubScalar, GENERATOR_EXTENDED, GENERATOR_NUMS_EXTENDED,
+};
+
+/// Use the pedersen commitment scheme to compute a transparent value
+/// commitment.
+pub fn transparent_value_commitment(value: u64) -> JubJubAffine {
+    JubJubAffine::from(GENERATOR_EXTENDED * JubJubScalar::from(value))
+}
+
+/// Use the pedersen commitment scheme to compute a value commitment using a
+/// blinding-factor.
+pub fn value_commitment(
+    value: u64,
+    blinding_factor: JubJubScalar,
+) -> JubJubAffine {
+    JubJubAffine::from(
+        (GENERATOR_EXTENDED * JubJubScalar::from(value))
+            + (GENERATOR_NUMS_EXTENDED * blinding_factor),
+    )
+}